requirements.txt
requests==2.31.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
pandas==1.5.3
schedule==1.2.1 